Further tune sqlite compile options
When compiled with SQLITE_THREADSAFE=2, SQLite can be used in a
multithreaded program so long as no two threads attempt to use the
same database connection (or any prepared statements derived from that
database connection) at the same time.

This can be safe than the default SQLITE_THREADSAFE=1 (Serialized),
but because this library uses a connection pool the constraints are
always maintained.